refactor: Add AddRecentSearchUseCase to SearchViewModel and update BaseViewModel
The `SearchViewModel` now includes the `AddRecentSearchUseCase` for managing recent searches.
The `BaseViewModel`'s `tryToExecute` method's `onSuccess` parameter has been updated to be a suspend function.

Additionally, a temporary log and a TODO comment have been added to `loadRecentSearches` in `SearchViewModel` for testing purposes and a TODO comment for an image resource in `SearchResultContent`.