Add internet and network state permissions
This commit adds the `INTERNET` and `ACCESS_NETWORK_STATE` permissions to the following AndroidManifest.xml files:
- `datasource/remote/search/src/main/AndroidManifest.xml`
- `app/src/main/AndroidManifest.xml`
- `repository/search/src/main/java/com/repository/search/AndroidManifest.xml`

Additionally, it reorders the operations in `SearchMediaRepositoryImpl` to add the search query to the history *before* adding the media to the local data source. This ensures that the search query is saved even if adding the media fails.