refactor: Change ContentScale from Crop to FillBounds in AsyncImage
The `contentScale` property for `AsyncImage` in `CompanyProductionCard.kt` and `GallerySection.kt` has been updated from `ContentScale.Crop` to `ContentScale.FillBounds`.

This change ensures that images will scale to fill the bounds of the `AsyncImage` composable, potentially distorting the aspect ratio if the image and bounds aspect ratios do not match. Previously, `ContentScale.Crop` would scale the image to fill the bounds while maintaining the aspect ratio, cropping any parts of the image that extended beyond the bounds.