Refactor: Centralize language update logic in Activities
This commit introduces a common approach for handling language updates across multiple Activities. The `attachBaseContext` method is overridden in the following Activities to load the saved language preference from SharedPreferences and apply it to the Activity's context:

- `MediaDetailsActivity`
- `AuthActivity`
- `TopRatingActivity`
- `MainActivity` (app module)
- `ContinueWatchingActivity`
- `SearchActivity`
- `MainActivity` (onboardingUi module)

A private `updateLocale` function has been added to each of these Activities to encapsulate the logic for creating a new configuration with the selected locale and updating the context. This ensures consistent language application at the Activity level.