Refactor: Update MovieMapper to use Genre enum
This commit refactors the `MovieMapper.kt` to utilize the `Genre` enum instead of creating new `Genre` objects.

The `toEntity()` extension functions for `MovieGenreDto` and `GenreEntity` have been updated to find the corresponding `Genre` enum entry based on the `name` property. If no matching entry is found, it defaults to `Genre.Unknown`.