Feat: Add genre resource mapper
This commit introduces a new `GenreResourceMapper` object in the `searchUi` module. This mapper provides a way to get a drawable resource ID for a given `Genre` enum value.

The `GenreResourceMapper` contains a map of `Genre` enum values to their corresponding drawable resource IDs. It also includes a `getResourceId` function that takes a `Genre` as input and returns the associated resource ID. If a genre is not found in the map, it defaults to `R.drawable.ic_category_all`.