Feat: Add VideoPlayer to MovieDetailsScreen
This commit introduces the `VideoPlayer` composable to the `MovieDetailsScreen`.

The `VideoPlayer` is displayed when `state.movieDetailsUiState.isYoutubePlayerVisible` is true and `state.movieDetailsUiState.youtubeVideoKey` is not null or empty. It allows users to watch a YouTube video related to the movie.

The `VideoPlayer` takes the `videoKey` and an `onCloseClick` lambda as parameters. The `onCloseClick` lambda is triggered when the user clicks the close button on the player, which then calls `movieDetailsScreenInteractionListener.closeYoutubePlayer()` to hide the player.