Add ActorPopularityRepository and its implementation
This commit introduces the `ActorPopularityRepository` interface in the `domain/guessGame` module and its implementation `ActorPopularityRepositoryImpl` in the `repository/guessGame` module.

The `ActorPopularityRepository` defines a suspend function `getPopularActor()` to fetch a list of popular actors.

The `ActorPopularityRepositoryImpl` implements this interface, utilizing a `NetworkConnectionChecker`, `ActorPopularityRemoteDataSource`, and `SettingLocalDataSource`. It fetches popular actors based on the user's selected language and handles potential network or API errors with a `safeCall` utility function.