Refactor: Add `@Serializable` to `UiGameLevel` and `QuestionType` enums
Added the `@Serializable` annotation to the `UiGameLevel` and `QuestionType` enums in the guess game feature. This allows these enums to be serialized and deserialized, which is necessary for passing them as navigation arguments.

The `@Keep` annotation was removed from `QuestionType` as it is no longer needed with the addition of `@Serializable` and corresponding Proguard rules.

Updated Proguard rules to keep `UiGameLevel` and `QuestionType` classes during minification.